move param prompt modifications into their own file
