+++ demo/index.html
@@ -29,6 +29,7 @@
     font-family: sans-serif;
     background-color: #eee;
     align-items: flex-end;
+    cursor: default;
   }
 
   .prosemirror-prompt-row-label {
